% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.progressBar.R
\name{print.progressBar}
\alias{print.progressBar}
\title{progressBar printing function}
\usage{
\method{print}{progressBar}(progress, ...)
}
\arguments{
\item{progress}{}
}
\description{
progressBar printing function
}
\seealso{
\code{\link{progressBar}}
}
