% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressBar.R
\name{progressBar}
\alias{progressBar}
\title{Set progressbar}
\usage{
progressBar(value, stroke.width = 1, easing = "easeInOut",
  duration = 3500, trail.color = "#eee", trail.width = 1,
  color = "#ffeb3b", svg.width = "100\%", svg.height = "100\%",
  color.from = NULL, color.to = NULL, width = NULL, height = 8,
  elementId = NULL)
}
\arguments{
\item{value}{percentage: to pass less than \code{1}.}

\item{stroke.width}{width of stroke.}

\item{easing}{easing function.}

\item{duration}{animation duration.}

\item{trail.color}{color of trail.}

\item{trail.width}{width of trail.}

\item{color}{color of bar.}

\item{svg.width, svg.height}{width and height of svg.}

\item{color.from, color.to}{starting and ending color (optional) see examples.}

\item{width, height}{dimensions of container.}

\item{elementId}{id of container.}
}
\description{
Create a progressbar
}
