% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{type}
\alias{type}
\alias{progressLine}
\alias{progressCircle}
\alias{progressSemi}
\title{Set type}
\usage{
progressLine(p)

progressCircle(p)

progressSemi(p)
}
\arguments{
\item{p}{an object of class \code{progressBar} as returned by \code{link{progressBar}}}
}
\description{
Change default type of progress bar from \code{line} to \code{circle} or \code{semi-circle}
}
